[master] Misc improvements (#3346)
* Use constexpr in SafeMath.
* Move AddressConversion to AccountData.
* Remove unused files & code.
* Remove unused temp database.
* Parallelize with half the cores to prevent the machine/docker from choking.
* Fixes for AppleClang not supporting some features of ranges yet
