+++ src/depends/common/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_library (Common SHARED CommonData.cpp CommonIO.cpp FileSystem.cpp FixedHash.cpp RLP.cpp SHA3.cpp Miner.cpp)
+add_library (Common SHARED CommonData.cpp FixedHash.cpp RLP.cpp SHA3.cpp Miner.cpp)
 include_directories(${Boost_INCLUDE_DIRS})
 target_include_directories (Common PUBLIC ${PROJECT_SOURCE_DIR}/src)
-target_link_libraries (Common PUBLIC Boost::filesystem Boost::system g3log)
+target_link_libraries (Common PUBLIC Boost::system g3log)
