+++ src/libBlockchain/DSBlock.cpp
@@ -38,7 +38,11 @@ bool SetDSBlock(zbytes& dst, const unsigned int offset,
   return SerializeToArray(result, dst, offset);
 }
 
+#ifdef __APPLE__
+template <typename RangeT>
+#else
 template <std::ranges::contiguous_range RangeT>
+#endif
 bool GetDSBlock(RangeT&& src, unsigned int offset, DSBlock& dsBlock) {
   if (offset >= src.size()) {
     LOG_GENERAL(WARNING, "Invalid data and offset, data size "
