+++ src/libBlockchain/DSBlockHeader.cpp
@@ -38,7 +38,11 @@ bool SetDSBlockHeader(zbytes& dst, unsigned int offset,
   return SerializeToArray(result, dst, offset);
 }
 
+#ifdef __APPLE__
+template <typename RangeT>
+#else
 template <std::ranges::contiguous_range RangeT>
+#endif
 bool GetDSBlockHeader(RangeT&& src, unsigned int offset,
                       DSBlockHeader& dsBlockHeader) {
   if (offset >= src.size()) {
