+++ src/libBlockchain/TxBlock.cpp
@@ -38,7 +38,11 @@ bool SetTxBlock(zbytes& dst, const unsigned int offset,
   return SerializeToArray(result, dst, offset);
 }
 
+#ifdef __APPLE__
+template <typename RangeT>
+#else
 template <std::ranges::contiguous_range RangeT>
+#endif
 bool GetTxBlock(RangeT&& src, unsigned int offset, TxBlock& txBlock) {
   if (offset >= src.size()) {
     LOG_GENERAL(WARNING, "Invalid data and offset, data size "
