+++ src/libBlockchain/VCBlock.cpp
@@ -39,7 +39,11 @@ bool SetVCBlock(zbytes& dst, const unsigned int offset,
   return SerializeToArray(result, dst, offset);
 }
 
+#ifdef __APPLE__
+template <typename RangeT>
+#else
 template <std::ranges::contiguous_range RangeT>
+#endif
 bool GetVCBlock(RangeT&& src, unsigned int offset, VCBlock& vcBlock) {
   if (offset >= src.size()) {
     LOG_GENERAL(WARNING, "Invalid data and offset, data size "
