+++ src/libData/AccountStore/AccountStore.cpp
@@ -361,32 +361,6 @@ bool AccountStore::IsPurgeRunning() {
           ContractStorage::GetContractStorage().IsPurgeRunning());
 }
 
-bool AccountStore::UpdateStateTrieFromTempStateDB() {
-  LOG_MARKER();
-
-  leveldb::Iterator* iter = nullptr;
-
-  while (iter == nullptr || iter->Valid()) {
-    vector<StateSharedPtr> states;
-    if (!BlockStorage::GetBlockStorage().GetTempStateInBatch(iter, states)) {
-      LOG_GENERAL(WARNING, "GetTempStateInBatch failed");
-      delete iter;
-      return false;
-    }
-    for (const auto& state : states) {
-      UpdateStateTrie(state->first, state->second);
-    }
-  }
-
-  delete iter;
-
-  if (!BlockStorage::GetBlockStorage().ResetDB(BlockStorage::TEMP_STATE)) {
-    LOG_GENERAL(WARNING, "BlockStorage::ResetDB (TEMP_STATE) failed");
-    return false;
-  }
-  return true;
-}
-
 bool AccountStore::RetrieveFromDisk() {
   InitSoft();
