+++ src/libData/AccountStore/AccountStore.h
@@ -106,9 +106,6 @@ class AccountStore  : public AccountStoreTrie {
   /// Reset the reference to underlying leveldb
   bool RefreshDB();
 
-  /// Use the states in Temp State DB to refresh the state merkle trie
-  bool UpdateStateTrieFromTempStateDB();
-
   /// commit the in-memory states into persistent storage
   bool MoveUpdatesToDisk(uint64_t dsBlockNum = 0);
