+++ src/libPersistence/BlockStorage.h
@@ -32,7 +32,6 @@ typedef std::tuple<uint32_t, uint64_t, uint64_t, BlockType, BlockHash>
     BlockLink;
 
 class TransactionWithReceipt;
-class Account;
 class LevelDB;
 
 typedef std::shared_ptr<DSBlock> DSBlockSharedPtr;
@@ -41,7 +40,6 @@ typedef std::shared_ptr<VCBlock> VCBlockSharedPtr;
 typedef std::shared_ptr<BlockLink> BlockLinkSharedPtr;
 typedef std::shared_ptr<MicroBlock> MicroBlockSharedPtr;
 typedef std::shared_ptr<TransactionWithReceipt> TxBodySharedPtr;
-typedef std::shared_ptr<std::pair<Address, Account>> StateSharedPtr;
 
 struct DiagnosticDataNodes {
   DequeOfShard shards;
@@ -96,7 +94,6 @@ class BlockStorage : boost::noncopyable {
   std::shared_ptr<LevelDB> m_blockLinkDB;
   std::shared_ptr<LevelDB> m_shardStructureDB;
   std::shared_ptr<LevelDB> m_stateDeltaDB;
-  std::shared_ptr<LevelDB> m_tempStateDB;
   std::shared_ptr<LevelDB> m_processedTxnTmpDB;
   // m_diagnosticDBNodes is needed only for LOOKUP_NODE_MODE, but to make the
   // unit test and monitoring tools work with the default setting of
@@ -131,7 +128,6 @@ class BlockStorage : boost::noncopyable {
     BLOCKLINK,
     SHARD_STRUCTURE,
     STATE_DELTA,
-    TEMP_STATE,
     DIAGNOSTIC_NODES,
     DIAGNOSTIC_COINBASE,
     STATE_ROOT,
@@ -282,13 +278,6 @@ class BlockStorage : boost::noncopyable {
 
   /// Retrieve state delta
   bool GetStateDelta(const uint64_t& finalBlockNum, zbytes& stateDelta);
-
-  /// Write state to tempState in batch
-  bool PutTempState(const std::unordered_map<Address, Account>& states);
-
-  /// Get state from tempState in batch
-  bool GetTempStateInBatch(leveldb::Iterator*& iter,
-                           std::vector<StateSharedPtr>& states);
 
   /// Save data for diagnostic / monitoring purposes (nodes in network)
   bool PutDiagnosticDataNodes(const uint64_t& dsBlockNum,
