+++ tests/Data/Test_Contract.cpp
@@ -30,7 +30,6 @@
 #include <Schnorr.h>
 #include "common/Constants.h"
 #include "common/TxnStatus.h"
-#include "depends/common/CommonIO.h"
 #include "libData/AccountData/Account.h"
 #include "libData/AccountData/Transaction.h"
 #include "libData/AccountData/TransactionReceipt.h"
