+++ tests/Data/Test_ContractInvoke.cpp
@@ -23,7 +23,6 @@
 #include <string>
 #include <vector>
 #include "common/Constants.h"
-#include "depends/common/CommonIO.h"
 #include "libData/AccountData/Account.h"
 #include "libData/AccountData/Transaction.h"
 #include "libData/AccountData/TransactionReceipt.h"
