+++ tests/depends/Trie/Test_TxnTrie.cpp
@@ -25,7 +25,6 @@
 #include <boost/filesystem/path.hpp>
 #include <boost/test/included/unit_test.hpp>
 
-#include "depends/common/CommonIO.h"
 #include "depends/common/FixedHash.h"
 #include "depends/json_spirit/JsonSpiritHeaders.h"
 #include "depends/libDatabase/MemoryDB.h"
